dex v2.26.0
This is so we can use dex's userinfo endpoint ([introduced](dexidp/dex@8b4dbb9) in v2.17).

It also changes dex's db volume to `emptyDir` type so it doesn't require any special config for Kind to create the dir in the host. the trade-off was that now dex stord data id truly ephemeral. If the pod goes down, all issued access tokens should be lost. This is obviously not good for production but enough for the demo.