Reconciliation of deleted (API key) secrets
Since the controller cannot find secrets that have been deleted, all it has to trigger reconciliation of affected services is the namespaced name of the deleted secret. So in this case of reconciliation of delete secrets, Authorino will dig into the cache of services looking for any service whose API key identity config has cached the deleted secret as an "authorized credential". These services will be reconciled.

For the already supported cases (create secret and update secret, the only change is that service reconciliation is now triggered inside a goroutine.