Auth metadata by generic HTTP request
Fecthes auth metadata from external HTTP service by GET or POST requests.

When POST is used, the resolved identity object (as JSON) is passed in the body of the request.

A shared secret between Authorino and the external HTTP service is expected to be defined for the service to authenticate the origin (`sharedSecretRef` property). The location where such secret value will travel in the request performed by Authorino to the HTTP service can be specified in a typical "credentials" property.