Change default authorino image to quay.io/3scale/authorino:latest
For the make targets `local-setup` and `local-rollout`, it will continue to build the image and push it to the local Kubernetes cluster started with Kind – now with the locally built image tagged as `authorino:local`.

This step of building an image locally can now be skept by supplying the `SKIP_LOCAL_BUILD=1` argument to the make `local-*` commands. In this case, the deployment defaults again to `AUTHORINO_IMAGE=quay.io/3scale/authorino:latest` and, just like with the more generic `make deploy`, Kubernetes will fetch the image from quay.

As usual, the `AUTHORINO_IMAGE` argument can still be used to set a different image, but it has no effect on the `local-*` make commands when `SKIP_LOCAL_BUILD=0` or unset (the default).