Make custom JSON properties to respect the original type of the source value instead of always converting to string
It works for Dynamic JSON properties as well as for Festival Wristband custom claims.

For static custom JSON property values (e.g. `spec.response.json.properties.value` or `spec.response.wristband.customClaims.value`), the JSON type of the property is the type of the static value as set (string, boolean, integer, array, object).

For values fetched from the authorization JSON (i.e. `valueFrom: { authJSON: <pattern> }`), the JSON type of the property is the type of the value fetched from the authorization JSON.

For example, the following spec:

```yaml
spec:
  response:
    - name: my-json-response
      json:
        properties:
          - name: prop1
            value: my-string
          - name: prop2
            value: true
          - name: prop3
            value: 123
          - name: prop3
            value: ["a", "b"]
          - name: prop5
            value: { c: 456 }
          - name: prop6
            valueFrom:
              authJSON: auth.identity.role # an array `["admin"]`
          - name: prop7
            valueFrom:
              authJSON: auth.identity.email_verified # the boolean value `false`
```

...generates the following JSON:

```jsonc
{
  "my-json-response": {
    "prop1": "my-string",
    "prop2": true,
    "prop3": 123,
    "prop4": [
      "a",
      "b"
    ],
    "prop5": {
      "c": 456
    },
    "prop6": [
      "admin"
    ],
    "prop7": false
  }
}
```