Cache of AuthConfigs with support for wildcard domains
It defines a Domain Search Tree (Domain Lookup Tree) based on a Radix Tree (trie) data structure.

It looks up for the longest match (longest common label) between the requested key (a host name) and the nodes of the tree, where each dot ('.') in the key is treated as a label separator that breaks down and induces a new level in the tree. Eventually, it looks up for wildcards in the tree, upwards from the leaf to the root, that can match the request and has an entry associated.