Add AuthConfig/Status conditions
+ summarises current state of linked hosts in Status.Summary.Hosts

An AuthConfig will only be considered ready (condition type `Ready`) when all host names from `spec.hosts` have been linked to the config in the cache. In cases where that is not possible due to a host name considered to be already taken (by other AuthConfig from a different k8s namespace), the AuthConfig will never be ready (condition reason `HostNotLinked`). Nevertheless, users can still rely on `status.summary.hostsReady: []string` for checking the state for other hosts that did succeed being linked to the config in the cache.

Summary status conditions and possible states:

| Type    | Status  | Reason          | Message                                      | Meaning |
|---------|---------|-----------------|----------------------------------------------|---------|
| `Ready` | `True`  | `Reconciled`    | N/A                                          | The AuthConfig is cached for all host names from `spec.hosts`. |
| `Ready` | `False` | `HostNotLinked` | One or more hosts not linked to the resource | Failed to cache the AuthConfig for one or more host names from `spec.hosts`. User should check `status.summary.hostsReady` for the list of host names that did succeed being linked to the AuthConfig in the cache; occasionally none. |

The counters and indicators in the status sub resource used to define the print columns when listing authconfig resources moved into Status.Summary. Moreover, 'Ready' is now an indicator of the number of hosts linked to the resource ('ready') in the format n/N, where n is number of hosts linked and N is the total number of hosts in spec.hosts. Column names were also modified for simplicity.