Bootstrap cache index on startup
This is needed when a new Authorino instance is started and finds pre-existing AuthConfigs marked as available (i.e. at least one host linked in the cache), so the new instance can respect the order of creation of the resources to populate its own cache from scratch. Otherwise, we risk incurring in issues such as the one described in https://github.com/Kuadrant/authorino/issues/317.