Reject linking AuthConfigs to hosts already taken regardless of namespace
Unifies the two traditional behaviours for `namespaced` reconciliation mode and for `cluster-wide` reconciliation mode regarding how a host is considered to be taken or not into one single behaviour. I.e., the namespace that the AuthConfigs belong to no longer matters; if a host is taken by an AuthConfig, it cannot be linked to another one.

Users who want to link a host to a different AuthConfig will first need to remove the host from the existing AuthConfig linked to the host and then add the new one. In the case of specific hosts vs. wildcard hosts, the more specific ones will always have to be linked before and then the more generic ones.