Embed the Authorino version info into the binary
The Authorino version info can be specified as a build argument (`VERSION`) of the `docker build` command, which is then set as an ldflag in the `go build` command.

When using the Makefile:
1. If `VERSION` is not specified, it defaults to the Git ref (SHA-1) of HEAD;
2. The image tag (`IMAGE_TAG`) is inferred from the value of `VERSION` (user-defined or set to default):
  - When it matches `^[0-9]+\.[0-9]+\.[0-9]+(-.+)?$` (e.g. '0.10.0-pre'), the tag is set to `v$VERSION` (e.g. 'v0.10.0-pre');
  - otherwise, the tag is set to fixed value 'local'.

Additionally, a new parameter `IMAGE_REPO` was introduced to the Makefile (default: 'authorino'), and the existing parameter `AUTHORINO_IMAGE` now defaults to `$(IMAGE_REPO):$(IMAGE_TAG)`.

In CI, `VERSION` is inferred based on `GITHUB_REF_NAME`:
- When the Git ref (branch/tag name) matches `^v[0-9]+\.[0-9]+\.[0-9]+(-.+)?$`, `VERSION` is set to `GITHUB_REF_NAME` (with all slashes replaced with dashes);
- otherwise, `VERSION` is set to the actual Git ref (i.e. `github.sha`)

Also in this chnage:
- New make target `docker-build`
- Rename the binary to 'authorino' (previously: 'manager')