Refactor: OPA.precompilePolicy no longer mutates the struct
This is for improved readability regarding the scope of the mu RWMutex. One who previously read the code could think that the precompilePolicy function was unsafe by not relying on the lock before touching opa.Rego (read) and opa.policy (write). This was not the case because precompilePolicy() was only used in the scope of OPA.updateRego(...), which does control the lock.

By ensuring precompilePolicy does not mutate the state of the struct, we not only make it clear that it doesn't need to use the lock, but also prevent further complications related to concurrency and the usage of this function, such as if ever invoked unsafely by other caller.